Finetuning Attempt
This was an attempt at fine-tuning llama-2-70b using Google Collab and Jupyter notebook. 
It was not very reliable on it's own, though with an embedded knowledge base, multiple SGPT/SLLM "Specialized GPT/LLMs" agents, and data retrieval access, a powerful tool to bring these Atomic/Subatomic interactions to light. 
The cost to compute is comparable to OpenAI and alternatives can be explored locally. 

For more information please visit www.conscious.energy